+++ pkg/apis/ml/v1alpha1/extension_spec_deployment.go
@@ -21,8 +21,8 @@
 package v1alpha1
 
 import (
+	schedulerPodApi "github.com/arangodb/kube-arangodb/pkg/apis/scheduler/v1alpha1/pod"
 	shared "github.com/arangodb/kube-arangodb/pkg/apis/shared"
-	sharedApi "github.com/arangodb/kube-arangodb/pkg/apis/shared/v1"
 	"github.com/arangodb/kube-arangodb/pkg/util"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 )
@@ -58,8 +58,8 @@ type ArangoMLExtensionSpecDeployment struct {
 	// Service defines how components will be exposed
 	Service *ArangoMLExtensionSpecDeploymentService `json:"service,omitempty"`
 
-	// PodTemplate defines base template for pods
-	*sharedApi.PodTemplate
+	// Pod defines base template for pods
+	*schedulerPodApi.Pod
 
 	// Prediction defines how Prediction workload will be deployed
 	Prediction *ArangoMLExtensionSpecDeploymentComponent `json:"prediction,omitempty"`
@@ -76,12 +76,12 @@ func (s *ArangoMLExtensionSpecDeployment) GetReplicas() int32 {
 	return *s.Replicas
 }
 
-func (s *ArangoMLExtensionSpecDeployment) GetPodTemplate() *sharedApi.PodTemplate {
-	if s == nil || s.PodTemplate == nil {
+func (s *ArangoMLExtensionSpecDeployment) GetPodTemplate() *schedulerPodApi.Pod {
+	if s == nil || s.Pod == nil {
 		return nil
 	}
 
-	return s.PodTemplate
+	return s.Pod
 }
 
 func (s *ArangoMLExtensionSpecDeployment) GetPrediction() *ArangoMLExtensionSpecDeploymentComponent {