+++ pkg/apis/ml/v1alpha1/storage_spec_test.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -27,6 +27,8 @@ import (
 	core "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 
+	schedulerContainerApi "github.com/arangodb/kube-arangodb/pkg/apis/scheduler/v1alpha1/container"
+	schedulerContainerResourcesApi "github.com/arangodb/kube-arangodb/pkg/apis/scheduler/v1alpha1/container/resources"
 	sharedApi "github.com/arangodb/kube-arangodb/pkg/apis/shared/v1"
 	"github.com/arangodb/kube-arangodb/pkg/util"
 )
@@ -66,8 +68,8 @@ func Test_ArangoMLStorageSpec(t *testing.T) {
 				core.ResourceMemory: resource.MustParse("128Mi"),
 			},
 		}
-		s.Mode.Sidecar.ContainerTemplate = &sharedApi.ContainerTemplate{}
-		s.Mode.Sidecar.Resources = &sharedApi.Resources{Resources: &assignedRequirements}
+		s.Mode.Sidecar.Container = &schedulerContainerApi.Container{}
+		s.Mode.Sidecar.Resources = &schedulerContainerResourcesApi.Resources{Resources: &assignedRequirements}
 
 		expectedRequirements := core.ResourceRequirements{
 			Requests: assignedRequirements.Requests,
@@ -77,7 +79,7 @@ func Test_ArangoMLStorageSpec(t *testing.T) {
 			},
 		}
 
-		actualRequirements := s.Mode.Sidecar.GetResources().With(&sharedApi.Resources{Resources: &core.ResourceRequirements{
+		actualRequirements := s.Mode.Sidecar.GetResources().With(&schedulerContainerResourcesApi.Resources{Resources: &core.ResourceRequirements{
 			Limits: core.ResourceList{
 				core.ResourceCPU:    resource.MustParse("100m"),
 				core.ResourceMemory: resource.MustParse("128Mi"),
@@ -87,6 +89,6 @@ func Test_ArangoMLStorageSpec(t *testing.T) {
 				core.ResourceMemory: resource.MustParse("256Mi"),
 			},
 		}})
-		require.Equal(t, expectedRequirements, *actualRequirements.GetResources())
+		require.Equal(t, expectedRequirements, actualRequirements.GetResources())
 	})
 }