+++ pkg/deployment/reconcile/action_runtime_container_image_update.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/deployment/rotation"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/container"
 )
 
 func newRuntimeContainerImageUpdateAction(action api.Action, actionCtx ActionContext) Action {
@@ -285,13 +286,13 @@ func (a actionRuntimeContainerImageUpdate) CheckProgress(ctx context.Context) (b
 		return true, false, nil
 	}
 
-	cspec, ok := k8sutil.GetContainerByName(pod, name)
+	cspec, ok := container.GetContainerByName(pod, name)
 	if !ok {
 		a.log.Info("Unable to find container spec")
 		return true, false, nil
 	}
 
-	cstatus, ok := k8sutil.GetContainerStatusByName(pod, name)
+	cstatus, ok := container.GetContainerStatusByName(pod, name)
 	if !ok {
 		a.log.Info("Unable to find container status")
 		return true, false, nil