+++ pkg/util/refs.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -44,6 +44,15 @@ func TypeOrDefault[T interface{}](input *T, defaultValue ...T) T {
 		return def
 	}
 	return *input
+}
+
+// WithDefault returns generic default value for type T if in is nil
+func WithDefault[T interface{}](in *T) T {
+	if in == nil {
+		return Default[T]()
+	}
+
+	return *in
 }
 
 // Default returns generic default value for type T